requirements.txt
amqp==1.4.9
anyjson==0.3.3
awesome-slugify==1.6.2
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
boto==2.38.0
celery==3.1.20
-e git+https://github.com/Gpetrak/crete-gis@b024a9d01434eb9699a7df75b6f08c80249651ff#egg=crete_gis
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.8.7
django-activity-stream==0.6.4
django-appconf==1.0.2
django-autocomplete-light==2.3.3
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.11.0
django-celery==3.1.17
django-cors-headers==2.1.0
django-downloadview==1.8
django-extensions==1.6.1
django-floppyforms==1.7.0
django-forms-bootstrap==3.1.0
django-friendly-tag-loader==1.2.1
django-geoexplorer==4.0.10
django-geojson==2.10.0
django-geonode-client==0.0.15
django-guardian==1.4.6
django-haystack==2.6.0
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
django-leaflet==0.22.0
django-location-field==2.0.3
django-modeltranslation==0.12
django-mptt==0.8.6
django-nose==1.4.4
django-oauth-toolkit==0.12.0
django-pagination==1.0.7
django-polymorphic==1.3
django-storages==1.6.5
django-taggit==0.21.0
django-tastypie==0.13.0
django-toolbelt==0.0.1
django-treebeard==4.0
django-user-accounts==2.0.2.dev0
dnspython==1.15.0
elasticsearch==2.4.0
email-validator==1.0.2
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.3
flake8==2.5.4
geolinks==0.2.0
-e git+https://github.com/GeoNode/geonode@5004d8772564ce8076db0fe34d32f667c38a0dcc#egg=GeoNode
geonode-agon-ratings==0.3.5
geonode-announcements==1.0.8
geonode-arcrest==10.2
geonode-avatar==2.1.6
geonode-dialogos==0.7
geonode-user-messages==0.1.6
gisdata==0.5.4
gsconfig==1.0.6
gsimporter==1.0.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
kombu==3.0.35
lxml==3.6.2
mccabe==0.4.0
MultipartPostHandler==0.1.0
nose==1.3.7
oauthlib==2.0.2
OWSLib==0.14.dev0
pathlib2==2.3.0
Paver==1.2.4
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
pinax-notifications==3.0.1
pinax-theme-bootstrap==3.0a11
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pycsw==2.0.3
pyelasticsearch==0.6.1
pyflakes==1.6.0
Pygments==2.2.0
pyproj==1.9.5.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.6.1
PyYAML==3.11
redis==2.10.5
regex==2016.7.21
requests==2.11.1
scandir==1.5
Shapely==1.5.17
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
static3==0.7.0
traitlets==4.3.2
transifex-client==0.12.3
Unidecode==0.4.19
urllib3==1.22
wcwidth==0.1.7
xlrd==1.0.0
xmltodict==0.9.2
